+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/PlaybackInfo.java
@@ -15,6 +15,7 @@
  */
 package androidx.media3.exoplayer;
 
+import android.os.SystemClock;
 import androidx.annotation.CheckResult;
 import androidx.annotation.Nullable;
 import androidx.media3.common.C;
@@ -23,6 +24,7 @@ import androidx.media3.common.PlaybackParameters;
 import androidx.media3.common.Player;
 import androidx.media3.common.Player.PlaybackSuppressionReason;
 import androidx.media3.common.Timeline;
+import androidx.media3.common.util.Util;
 import androidx.media3.exoplayer.source.MediaSource.MediaPeriodId;
 import androidx.media3.exoplayer.source.TrackGroupArray;
 import androidx.media3.exoplayer.trackselection.TrackSelectorResult;
@@ -92,6 +94,11 @@ import java.util.List;
    * in the {@link #timeline}, in microseconds.
    */
   public volatile long positionUs;
+  /**
+   * The value of {@link SystemClock#elapsedRealtime()} when {@link #positionUs} was updated, in
+   * milliseconds.
+   */
+  public volatile long positionUpdateTimeMs;
 
   /**
    * Creates an empty placeholder playback info which can be used for masking as long as no real
@@ -120,6 +127,7 @@ import java.util.List;
         /* bufferedPositionUs= */ 0,
         /* totalBufferedDurationUs= */ 0,
         /* positionUs= */ 0,
+        /* positionUpdateTimeMs= */ 0,
         /* sleepingForOffload= */ false);
   }
 
@@ -142,6 +150,7 @@ import java.util.List;
    * @param bufferedPositionUs See {@link #bufferedPositionUs}.
    * @param totalBufferedDurationUs See {@link #totalBufferedDurationUs}.
    * @param positionUs See {@link #positionUs}.
+   * @param positionUpdateTimeMs See {@link #positionUpdateTimeMs}.
    * @param sleepingForOffload See {@link #sleepingForOffload}.
    */
   public PlaybackInfo(
@@ -162,6 +171,7 @@ import java.util.List;
       long bufferedPositionUs,
       long totalBufferedDurationUs,
       long positionUs,
+      long positionUpdateTimeMs,
       boolean sleepingForOffload) {
     this.timeline = timeline;
     this.periodId = periodId;
@@ -180,6 +190,7 @@ import java.util.List;
     this.bufferedPositionUs = bufferedPositionUs;
     this.totalBufferedDurationUs = totalBufferedDurationUs;
     this.positionUs = positionUs;
+    this.positionUpdateTimeMs = positionUpdateTimeMs;
     this.sleepingForOffload = sleepingForOffload;
   }
 
@@ -231,6 +242,7 @@ import java.util.List;
         bufferedPositionUs,
         totalBufferedDurationUs,
         positionUs,
+        /* positionUpdateTimeMs= */ SystemClock.elapsedRealtime(),
         sleepingForOffload);
   }
 
@@ -260,6 +272,7 @@ import java.util.List;
         bufferedPositionUs,
         totalBufferedDurationUs,
         positionUs,
+        positionUpdateTimeMs,
         sleepingForOffload);
   }
 
@@ -289,6 +302,7 @@ import java.util.List;
         bufferedPositionUs,
         totalBufferedDurationUs,
         positionUs,
+        positionUpdateTimeMs,
         sleepingForOffload);
   }
 
@@ -318,6 +332,7 @@ import java.util.List;
         bufferedPositionUs,
         totalBufferedDurationUs,
         positionUs,
+        positionUpdateTimeMs,
         sleepingForOffload);
   }
 
@@ -347,6 +362,7 @@ import java.util.List;
         bufferedPositionUs,
         totalBufferedDurationUs,
         positionUs,
+        positionUpdateTimeMs,
         sleepingForOffload);
   }
 
@@ -376,6 +392,7 @@ import java.util.List;
         bufferedPositionUs,
         totalBufferedDurationUs,
         positionUs,
+        positionUpdateTimeMs,
         sleepingForOffload);
   }
 
@@ -409,6 +426,7 @@ import java.util.List;
         bufferedPositionUs,
         totalBufferedDurationUs,
         positionUs,
+        positionUpdateTimeMs,
         sleepingForOffload);
   }
 
@@ -438,6 +456,7 @@ import java.util.List;
         bufferedPositionUs,
         totalBufferedDurationUs,
         positionUs,
+        positionUpdateTimeMs,
         sleepingForOffload);
   }
 
@@ -467,6 +486,99 @@ import java.util.List;
         bufferedPositionUs,
         totalBufferedDurationUs,
         positionUs,
+        positionUpdateTimeMs,
         sleepingForOffload);
+  }
+
+  /**
+   * Copies playback info with new estimated playing position.
+   *
+   * <p>Position is estimated with {@link #positionUs}, {@link #positionUpdateTimeMs}, and {@link
+   * PlaybackParameters#speed}.
+   *
+   * @return Copied playback info with new, estimated playback position.
+   */
+  @CheckResult
+  public PlaybackInfo copyWithEstimatedPosition() {
+    return new PlaybackInfo(
+        timeline,
+        periodId,
+        requestedContentPositionUs,
+        discontinuityStartPositionUs,
+        playbackState,
+        playbackError,
+        isLoading,
+        trackGroups,
+        trackSelectorResult,
+        staticMetadata,
+        loadingMediaPeriodId,
+        playWhenReady,
+        playbackSuppressionReason,
+        playbackParameters,
+        bufferedPositionUs,
+        totalBufferedDurationUs,
+        getEstimatedPositionUs(),
+        SystemClock.elapsedRealtime(),
+        sleepingForOffload);
+  }
+
+  /**
+   * Sets new playing position with update time of {@link SystemClock#elapsedRealtime()}, time
+   * relative to the start of the associated period in the {@link #timeline}
+   *
+   * @param positionUs The new playing position.
+   */
+  public void updatePositionUs(long positionUs) {
+    // Write order of positionUs then positionUpdateTimeMs in order to be reverse of
+    // retrieval in getExtrapolatedPositionUs().
+    this.positionUs = positionUs;
+    this.positionUpdateTimeMs = SystemClock.elapsedRealtime();
+  }
+
+  /**
+   * Retrieves estimated position based on {@link #positionUs}, {@link #positionUpdateTimeMs}, and
+   * {@link PlaybackParameters#speed}.
+   *
+   * <p>If not playing, then the estimated position is {@link #positionUs}.
+   *
+   * @return The estimated position.
+   */
+  public long getEstimatedPositionUs() {
+    if (!isPlaying()) {
+      return this.positionUs;
+    }
+
+    // Snapshot of volatile position info
+    long positionUs;
+    long positionUpdateTimeMs;
+    do {
+      // Read order of positionUpdateTimeMs then positionUs to be reverse of updatePositionUs write.
+      positionUpdateTimeMs = this.positionUpdateTimeMs;
+      positionUs = this.positionUs;
+    } while (positionUpdateTimeMs != this.positionUpdateTimeMs);
+
+    long elapsedTimeMs = SystemClock.elapsedRealtime() - positionUpdateTimeMs;
+    long estimatedPositionMs =
+        Util.usToMs(positionUs) + (long) (elapsedTimeMs * playbackParameters.speed);
+    return Util.msToUs(estimatedPositionMs);
+  }
+
+  /**
+   * Returns whether this object represents a playing state.
+   *
+   * <p>Returns true if the following conditions are met:
+   *
+   * <ul>
+   *   <li>{@link #playbackState} is {@link Player#STATE_READY}
+   *   <li>{@link #playWhenReady} is true.
+   *   <li>{@link #playbackSuppressionReason} is {@link Player#PLAYBACK_SUPPRESSION_REASON_NONE}
+   * </ul>
+   *
+   * @return Whether the playbackInfo represents a playing state.
+   */
+  public boolean isPlaying() {
+    return playbackState == Player.STATE_READY
+        && playWhenReady
+        && playbackSuppressionReason == Player.PLAYBACK_SUPPRESSION_REASON_NONE;
   }
 }