+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/MediaPeriodQueueTest.java
@@ -1362,6 +1362,7 @@ public final class MediaPeriodQueueTest {
             /* bufferedPositionUs= */ 0,
             /* totalBufferedDurationUs= */ 0,
             /* positionUs= */ 0,
+            /* positionUpdateTimeMs= */ 0,
             /* sleepingForOffload= */ false);
   }